use command pattern for shop talents
